Initiate a selection

Initiate a selection

select_single(
  encodings = NULL,
  fields = NULL,
  init = NULL,
  nearest = FALSE,
  on = "click",
  clear = "dblclick",
  empty = "all",
  resolve = "global"
)

select_multi(
  encodings = NULL,
  fields = NULL,
  init = NULL,
  toggle = TRUE,
  nearest = FALSE,
  on = "click",
  clear = "dblclick",
  empty = "all",
  resolve = "global"
)

select_interval(
  encodings = c("x", "y"),
  init = NULL,
  mark = NULL,
  on = "[mousedown, window:mouseup] > window:mousemove!",
  clear = "dblclick",
  translate = on,
  empty = "all",
  zoom = TRUE,
  resolve = "global"
)

select_legend(fields, on = "click", clear = "dblclick")

select_domain()

select_bind(...)

Arguments

encodings
A character vector of encoding channels, such as "x" and "y".

fields
A character vector of data fields.

init
An initial value upon selection.

nearest
If FALSE, data values must be interacted with directly to be added to the selection.

on, clear
An event type that triggers/clears the selection. Options are "click", "dblclick", "dragenter", "dragleave", "dragover", "keydown", "keypress", "keyup", "mousedown", "mouseover", "mousemove", "mouseout", "mouseup", "mousewheel", "touchend", "touchmove", "touchstart", "wheel".

empty
An empty selection includes "all" or "none" data values.

resolve
One of "global", "union", "intersect" options to resolve ambiguity for layered and multi-view displays.

toggle
A logical to control whether data values should be toggled or only ever inserted into multi selections.

mark
A named vector of mark properties for brushed rectangle.

translate
A string or logical to interactively move an interval selection back-and-forth.

zoom
If TRUE, interactively resize an interval selection.

...
A set of name-value pairs with data variables on the LHS and input_*() on the RHS.

Composing Multiple Selections

A set of operations ...